+++ rust/src/api/types/booked.rs
@@ -7,13 +7,13 @@ use std::{
 };
 use time::Date;
 
-use crate::api::types::{ElementIdx, iso8601date, raw};
+use crate::api::types::{iso8601date, raw, ElementIdx};
 
 /// A Beancount directive of a particular [DirectiveVariant].
 #[derive(Serialize, PartialEq, Eq, Clone, Debug)]
 #[serde(rename_all = "kebab-case")]
 pub struct Directive<'a> {
-    pub(crate) raw_idx: usize,
+    pub(crate) idx: usize,
     #[serde(with = "iso8601date")]
     pub(crate) date: Date,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -66,7 +66,7 @@ pub struct Transaction<'a> {
 #[serde(rename_all = "kebab-case")]
 pub(crate) struct Posting<'a> {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) raw_idx: Option<usize>,
+    pub(crate) idx: Option<usize>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub(crate) flag: Option<Cow<'static, str>>,
     pub(crate) acc: &'a str,